Use pgkutil instead of importlib_resources

+++ scripts/dependency_test.sh
@@ -52,7 +52,9 @@ test_service() {
   pip install -r requirements-tests.txt > /dev/null
   pip install .[$service] > /dev/null 2>&1
   pip install boto > /dev/null 2>&1
-  pip uninstall setuptools pkg_resources -y > /dev/null 2>&1
+  if [[ $service != "xray" ]]; then
+    pip uninstall setuptools pkg_resources -y > /dev/null 2>&1
+  fi
   # Restart venv - ensure these deps are loaded
   deactivate
   source ${venv_path}/bin/activate > /dev/null
